cosmos_django/requirements.txt
asgiref==3.2.7
astroid==2.4.1
cffi==1.14.0
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.3.0
django-rest-knox==4.1.0
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.2.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.770
mypy-extensions==0.4.3
Pillow==7.1.2
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
pytz==2020.1
selenium==3.141.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wrapt==1.12.1
yapf==0.30.0
